Issue #408 :
  Refactored and moved Insertion form to InserisciFestivitaForm.js
